+++ Backend/Web_API/DTOs/Response/ProfileResponseDto.cs
@@ -2,7 +2,7 @@
 
 public record ProfileResponseDto
 {
-    public int ProfileId { get; init; } = 0;
+    public long ProfileId { get; init; } = 0;
     public string UserName { get; init; } = "";
     public string FirstName { get; init; } = "";
     public string LastName { get; init; } = "";
